notifications/notification_consumer_to_all/requirements.txt
certifi==2020.12.5
Jinja2==3.0.1
MarkupSafe==2.0.1
notificator-yandprakt-middle-python==0.0.5
pika==1.2.0
psycopg2-binary==2.9.3
pydantic==1.8.1
python-logstash==0.4.6
typing-extensions==3.7.4.3
urllib3==1.26.4
